write interpreter tests 1

+++ src/main/java/basic/IntegerNode.java
@@ -7,7 +7,7 @@ public class IntegerNode extends Node {
         value = in;
     }
 
-    public float getValue() {
+    public int getValue() {
         return value;
     }
